+++ src/02-tagger/04-correction/matches/02-dates.js
@@ -81,12 +81,14 @@ let list = [
   { match: `(#Place|#Demonmym|#Time) (standard|daylight|central|mountain)? time`, tag: 'Timezone', reason: 'std-time' },
   // eastern time
   {
-    match: `(eastern|mountain|pacific|central) (standard|daylight|summer)? time`,
+    match: `(eastern|mountain|pacific|central|atlantic) (standard|daylight|summer)? time`,
     tag: 'Timezone',
     reason: 'eastern-time',
   },
   // 5pm central
   { match: `#Time [(eastern|mountain|pacific|central|est|pst|gmt)]`, group: 0, tag: 'Timezone', reason: '5pm-central' },
+  // central european time
+  { match: `(central|western|eastern) european time`, tag: 'Timezone', reason: 'cet' },
 ]
 
 module.exports = list